+++ snowball.go
@@ -3,24 +3,31 @@ package snowball
 import (
 	"fmt"
 	"github.com/kljensen/snowball/english"
+	"github.com/kljensen/snowball/french"
 	"github.com/kljensen/snowball/spanish"
 )
 
 const (
-	VERSION string = "v0.1.1"
+	VERSION string = "v0.2.0"
 )
 
 // Stem a word in the specified language.
 //
 func Stem(word, language string, stemStopWords bool) (stemmed string, err error) {
+
+	var f func(string, bool) string
 	switch language {
 	case "english":
-		stemmed = english.Stem(word, stemStopWords)
+		f = english.Stem
 	case "spanish":
-		stemmed = spanish.Stem(word, stemStopWords)
+		f = spanish.Stem
+	case "french":
+		f = french.Stem
 	default:
 		err = fmt.Errorf("Unknown language: %s", language)
+		return
 	}
+	stemmed = f(word, stemStopWords)
 	return
 
 }